backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
